chore: remove v1 public api routes, contracts, and related code
Remove the unused v1 public REST API that was never publicly launched:
- Delete all 17 v1 route handlers and 4 integration tests
- Delete public-api library (handler, auth, errors, request-id)
- Delete public API contracts from @vm0/core
- Remove public contract re-exports from contracts/index.ts
- Move pagination schemas from public/common to platform.ts
- Clean up 5 dead public API methods from CLI api-client
- Update middleware to remove /v1/ route classification
- Update test helpers to use internal API for getTestRun
- Delete E2E test file (public-api-v1.bats)
- Update knip config to remove v1 entry patterns

Closes #3389
