+++ turbo/apps/cli/src/lib/api/api-client.ts
@@ -20,11 +20,6 @@ import {
   schedulesByNameContract,
   schedulesEnableContract,
   scheduleRunsContract,
-  publicAgentsListContract,
-  publicArtifactsListContract,
-  publicArtifactByIdContract,
-  publicVolumesListContract,
-  publicVolumeByIdContract,
   credentialsMainContract,
   credentialsByNameContract,
   realtimeTokenContract,
@@ -34,11 +29,6 @@ import {
   type ScheduleListResponse,
   type DeployScheduleResponse,
   type ScheduleRunsResponse,
-  type PublicAgent,
-  type PublicArtifact,
-  type PublicArtifactDetail,
-  type PublicVolume,
-  type PublicVolumeDetail,
   type CredentialResponse,
   type CredentialListResponse,
   type AblyTokenRequest,
@@ -1008,145 +998,6 @@ class ApiClient {
     const message =
       errorBody.error?.message ||
       `Failed to list runs for schedule "${params.name}"`;
-    throw new Error(message);
-  }
-
-  /**
-   * List public agents
-   */
-  async listPublicAgents(query?: {
-    cursor?: string;
-    limit?: number;
-    name?: string;
-  }): Promise<{
-    data: PublicAgent[];
-    pagination: { nextCursor: string | null; hasMore: boolean };
-  }> {
-    const baseUrl = await this.getBaseUrl();
-    const headers = await this.getHeaders();
-
-    const client = initClient(publicAgentsListContract, {
-      baseUrl,
-      baseHeaders: headers,
-      jsonQuery: false,
-    });
-
-    const result = await client.list({ query: query ?? {} });
-
-    if (result.status === 200) {
-      return result.body;
-    }
-
-    const errorBody = result.body as ApiErrorResponse;
-    const message = errorBody.error?.message || "Failed to list agents";
-    throw new Error(message);
-  }
-
-  /**
-   * List public artifacts
-   */
-  async listPublicArtifacts(query?: {
-    cursor?: string;
-    limit?: number;
-  }): Promise<{
-    data: PublicArtifact[];
-    pagination: { nextCursor: string | null; hasMore: boolean };
-  }> {
-    const baseUrl = await this.getBaseUrl();
-    const headers = await this.getHeaders();
-
-    const client = initClient(publicArtifactsListContract, {
-      baseUrl,
-      baseHeaders: headers,
-      jsonQuery: false,
-    });
-
-    const result = await client.list({ query: query ?? {} });
-
-    if (result.status === 200) {
-      return result.body;
-    }
-
-    const errorBody = result.body as ApiErrorResponse;
-    const message = errorBody.error?.message || "Failed to list artifacts";
-    throw new Error(message);
-  }
-
-  /**
-   * Get public artifact by ID
-   */
-  async getPublicArtifact(id: string): Promise<PublicArtifactDetail> {
-    const baseUrl = await this.getBaseUrl();
-    const headers = await this.getHeaders();
-
-    const client = initClient(publicArtifactByIdContract, {
-      baseUrl,
-      baseHeaders: headers,
-      jsonQuery: false,
-    });
-
-    const result = await client.get({ params: { id } });
-
-    if (result.status === 200) {
-      return result.body;
-    }
-
-    const errorBody = result.body as ApiErrorResponse;
-    const message = errorBody.error?.message || `Artifact "${id}" not found`;
-    throw new Error(message);
-  }
-
-  /**
-   * List public volumes
-   */
-  async listPublicVolumes(query?: {
-    cursor?: string;
-    limit?: number;
-  }): Promise<{
-    data: PublicVolume[];
-    pagination: { nextCursor: string | null; hasMore: boolean };
-  }> {
-    const baseUrl = await this.getBaseUrl();
-    const headers = await this.getHeaders();
-
-    const client = initClient(publicVolumesListContract, {
-      baseUrl,
-      baseHeaders: headers,
-      jsonQuery: false,
-    });
-
-    const result = await client.list({ query: query ?? {} });
-
-    if (result.status === 200) {
-      return result.body;
-    }
-
-    const errorBody = result.body as ApiErrorResponse;
-    const message = errorBody.error?.message || "Failed to list volumes";
-    throw new Error(message);
-  }
-
-  /**
-   * Get public volume by ID
-   */
-  async getPublicVolume(id: string): Promise<PublicVolumeDetail> {
-    const baseUrl = await this.getBaseUrl();
-    const headers = await this.getHeaders();
-
-    const client = initClient(publicVolumeByIdContract, {
-      baseUrl,
-      baseHeaders: headers,
-      jsonQuery: false,
-    });
-
-    const result = await client.get({ params: { id } });
-
-    if (result.status === 200) {
-      return result.body;
-    }
-
-    const errorBody = result.body as ApiErrorResponse;
-    const message = errorBody.error?.message || `Volume "${id}" not found`;
     throw new Error(message);
   }
