--- turbo/apps/web/app/v1/agents/[id]/route.ts
@@ -1,94 +1,0 @@
-/**
- * Public API v1 - Agent by ID Endpoint
- *
- * GET /v1/agents/:id - Get agent details
- */
-import { initServices } from "../../../../src/lib/init-services";
-import {
-  createPublicApiHandler,
-  tsr,
-} from "../../../../src/lib/public-api/handler";
-import { publicAgentByIdContract } from "@vm0/core";
-import {
-  authenticatePublicApi,
-  isAuthSuccess,
-} from "../../../../src/lib/public-api/auth";
-import { resolveScope } from "../../../../src/lib/scope/resolve-scope";
-import { agentComposes } from "../../../../src/db/schema/agent-compose";
-import { eq, and } from "drizzle-orm";
-
-const router = tsr.router(publicAgentByIdContract, {
-  get: async ({ params, headers }) => {
-    initServices();
-
-    const auth = await authenticatePublicApi(headers.authorization);
-    if (!isAuthSuccess(auth)) {
-      return {
-        status: 401 as const,
-        body: {
-          error: {
-            type: "authentication_error" as const,
-            code: "invalid_api_key",
-            message: "Invalid API key provided",
-          },
-        },
-      };
-    }
-
-    // Get user's scope
-    const userScope = await resolveScope(auth.userId, headers.authorization);
-    if (!userScope) {
-      return {
-        status: 401 as const,
-        body: {
-          error: {
-            type: "authentication_error" as const,
-            code: "invalid_api_key",
-            message:
-              "Please set up your scope first. Login again with: vm0 login",
-          },
-        },
-      };
-    }
-
-    // Find agent by ID, ensuring it belongs to user's scope
-    const [agent] = await globalThis.services.db
-      .select()
-      .from(agentComposes)
-      .where(
-        and(
-          eq(agentComposes.id, params.id),
-          eq(agentComposes.scopeId, userScope.id),
-        ),
-      )
-      .limit(1);
-
-    if (!agent) {
-      return {
-        status: 404 as const,
-        body: {
-          error: {
-            type: "not_found_error" as const,
-            code: "resource_not_found",
-            message: `No such agent: '${params.id}'`,
-          },
-        },
-      };
-    }
-
-    return {
-      status: 200 as const,
-      body: {
-        id: agent.id,
-        name: agent.name,
-        currentVersionId: agent.headVersionId,
-        createdAt: agent.createdAt.toISOString(),
-        updatedAt: agent.updatedAt.toISOString(),
-      },
-    };
-  },
-});
-
-const handler = createPublicApiHandler(publicAgentByIdContract, router);
-
-export { handler as GET };
