--- turbo/apps/web/app/v1/runs/[id]/cancel/route.ts
@@ -1,188 +1,0 @@
-/**
- * Public API v1 - Cancel Run Endpoint
- *
- * POST /v1/runs/:id/cancel - Cancel a pending or running execution
- */
-import { initServices } from "../../../../../src/lib/init-services";
-import {
-  createPublicApiHandler,
-  tsr,
-} from "../../../../../src/lib/public-api/handler";
-import { publicRunCancelContract } from "@vm0/core";
-import {
-  authenticatePublicApi,
-  isAuthSuccess,
-} from "../../../../../src/lib/public-api/auth";
-import { agentRuns } from "../../../../../src/db/schema/agent-run";
-import {
-  agentComposes,
-  agentComposeVersions,
-} from "../../../../../src/db/schema/agent-compose";
-import { eq } from "drizzle-orm";
-import { killSandbox } from "../../../../../src/lib/sandbox/sandbox-service";
-
-interface RunResult {
-  output?: string;
-  checkpointId?: string;
-  agentSessionId?: string;
-  artifactName?: string;
-  artifactVersion?: string;
-  volumes?: Record<string, string>;
-}
-
-const CANCELLABLE_STATUSES = ["pending", "running"];
-
-function buildRunResponseBody(
-  run: typeof agentRuns.$inferSelect,
-  compose: typeof agentComposes.$inferSelect | null,
-  status:
-    | "pending"
-    | "running"
-    | "completed"
-    | "failed"
-    | "timeout"
-    | "cancelled",
-  error: string | null,
-) {
-  const runResult = run.result as RunResult | null;
-  let executionTimeMs: number | null = null;
-  if (run.startedAt && run.completedAt) {
-    executionTimeMs = run.completedAt.getTime() - run.startedAt.getTime();
-  }
-  return {
-    id: run.id,
-    agentId: compose?.id ?? "",
-    agentName: compose?.name ?? "unknown",
-    status,
-    prompt: run.prompt,
-    createdAt: run.createdAt.toISOString(),
-    startedAt: run.startedAt?.toISOString() ?? null,
-    completedAt: run.completedAt?.toISOString() ?? null,
-    error,
-    executionTimeMs,
-    checkpointId: runResult?.checkpointId ?? null,
-    sessionId: runResult?.agentSessionId ?? null,
-    artifactName: runResult?.artifactName ?? null,
-    artifactVersion: runResult?.artifactVersion ?? null,
-    volumes: runResult?.volumes,
-  };
-}
-
-const router = tsr.router(publicRunCancelContract, {
-  cancel: async ({ params, headers }) => {
-    initServices();
-
-    const auth = await authenticatePublicApi(headers.authorization);
-    if (!isAuthSuccess(auth)) {
-      return {
-        status: 401 as const,
-        body: {
-          error: {
-            type: "authentication_error" as const,
-            code: "invalid_api_key",
-            message: "Invalid API key provided",
-          },
-        },
-      };
-    }
-
-    // Find run by ID
-    const [result] = await globalThis.services.db
-      .select({
-        run: agentRuns,
-        compose: agentComposes,
-      })
-      .from(agentRuns)
-      .leftJoin(
-        agentComposeVersions,
-        eq(agentRuns.agentComposeVersionId, agentComposeVersions.id),
-      )
-      .leftJoin(
-        agentComposes,
-        eq(agentComposeVersions.composeId, agentComposes.id),
-      )
-      .where(eq(agentRuns.id, params.id))
-      .limit(1);
-
-    if (!result) {
-      return {
-        status: 404 as const,
-        body: {
-          error: {
-            type: "not_found_error" as const,
-            code: "resource_not_found",
-            message: `No such run: '${params.id}'`,
-          },
-        },
-      };
-    }
-
-    // Verify ownership
-    if (result.run.userId !== auth.userId) {
-      return {
-        status: 404 as const,
-        body: {
-          error: {
-            type: "not_found_error" as const,
-            code: "resource_not_found",
-            message: `No such run: '${params.id}'`,
-          },
-        },
-      };
-    }
-
-    const { run, compose } = result;
-
-    // Check if run is in a cancellable state
-    if (!CANCELLABLE_STATUSES.includes(run.status)) {
-      return {
-        status: 400 as const,
-        body: {
-          error: {
-            type: "invalid_request_error" as const,
-            code: "invalid_state",
-            message: `Run cannot be cancelled: current status is '${run.status}'`,
-          },
-        },
-      };
-    }
-
-    // Update run status to cancelled
-    const now = new Date();
-    const [updatedRun] = await globalThis.services.db
-      .update(agentRuns)
-      .set({
-        status: "cancelled",
-        completedAt: now,
-      })
-      .where(eq(agentRuns.id, params.id))
-      .returning();
-
-    if (!updatedRun) {
-      return {
-        status: 500 as const,
-        body: {
-          error: {
-            type: "api_error" as const,
-            code: "internal_error",
-            message: "Failed to cancel run",
-          },
-        },
-      };
-    }
-
-    // Kill E2B sandbox if it exists
-    if (run.sandboxId) {
-      await killSandbox(run.sandboxId);
-    }
-
-    return {
-      status: 200 as const,
-      body: buildRunResponseBody(updatedRun, compose, "cancelled", null),
-    };
-  },
-});
-
-const handler = createPublicApiHandler(publicRunCancelContract, router);
-
-export { handler as POST };
