+++ turbo/knip.json
@@ -6,7 +6,6 @@
         "app/**/page.{ts,tsx}",
         "app/**/layout.{ts,tsx}",
         "app/api/**/route.{ts,tsx}",
-        "app/v1/**/route.{ts,tsx}",
         "middleware.{ts,tsx}",
         "middleware.cors.ts",
         "i18n.ts",
@@ -113,7 +112,6 @@
     "apps/cli/src/lib/api/api-client.ts",
     "apps/cli/src/lib/utils/paginate.ts",
     "apps/cli/src/commands/cook/cook.ts",
-    "packages/core/src/contracts/public/agents.ts",
     "apps/web/src/lib/slack/**",
     "apps/web/src/lib/auth/sandbox-token.ts",
     "scripts/rebuild-snapshots.ts",
